Further speculative fix to the marquee.
Part of the problem seems to be that resuming the marquee during the slide transition won't work because it depends on the marquee being visible.  So, we schedule the onShow to run 250ms after the transition, by which time the marquee will be visible; the effect is still more-or-less invisible to the user.